Remove deprecated statement classes
