+++ UPGRADE.md
@@ -6,15 +6,20 @@
 - `DriverException`
 - `PDOConnection`
 - `PDOException`
+- `PDOStatement`
 - `IBMDB2\DB2Connection`
 - `IBMDB2\DB2Driver`
 - `IBMDB2\DB2Exception`
+- `IBMDB2\DB2Statement`
 - `Mysqli\MysqliConnection`
 - `Mysqli\MysqliException`
+- `Mysqli\MysqliStatement`
 - `OCI8\OCI8Connection`
 - `OCI8\OCI8Exception`
+- `OCI8\OCI8Statement`
 - `SQLSrv\SQLSrvConnection`
 - `SQLSrv\SQLSrvException`
+- `SQLSrv\SQLSrvStatement`
 
 ## BC BREAK: `ServerInfoAwareConnection::requiresQueryForServerVersion()` is removed.
